Index VariableCollections with __call__
This allows a convenient and consistent formulation
of models which work with or without indexed variables.

It also seems to make sense that Part.constraints() does not
work with several indices (e.g. different points in time) but
that it must be called once per time step. The code would otherwise
get confusing with *indices meaning both "dimensions of an index"
and "a list of (possibly multi-dimensional) indices".