Change ResourceNetwork and Cluster
Node objects are now responsible for their own balance constraints.
Clusters and Nodes know about each other: If a Node is in a Cluster,
it will not report balance constraints.

ResourceNetwork now only coordinates the connections, and owns the
flow variables. When an edge is added to a ResourceNetwork, the
network adds the flow variable to the corresponding nodes.

Also some general cleanup and simplification of the code.