Add constraints property to Variable
This simplifies handling of PiecewiseAffine:
PiecewiseAffine simply adds its constraints to all the
contained variables.

This has the effect that variables' constraints must
now be used in any optimization problem.

To be safe, and to clean things up, I now let the Problem object
report all its variables and constraints. It leads to a simplification
in the PulpEngine class.

Also, interface of Problem also changes because the constraints set
is now replaced by a property which computes the list of constraints.
A number of tests break.