Variables may not have Constraints
Because explicit is better than implicit! This simplifies things
quite a bit both implementation-wise and for making the library
transparent for the user.

PiecewiseAffine breaks, so I remove it. Piecewise affine functions need
to be implemented in some other way, probably by creating a helper
function to just generate the constraints. Not as pretty code, but much
more transparent.