Don't use variable names in PuLP
Saves a LOT of time.
(Could also save a lot of time by not having
such a stupid loop, but now I chose to just remove
it altogether and instead name variables in a sequence
for each problem. Harder to understand anything
from the produced problem files, but for now it does
not matter.)